pyproject.toml
[tool.poetry]
name = "lesson-27-2-hw"
version = "0.1.0"
description = ""
authors = ["mgd <mr.zizikk@gamil.com>"]
readme = "README.md"
packages = [{include = "lesson_27_2_hw"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
psycopg2-binary = "^2.9.5"
pillow = "^9.5.0"
djangorestframework = "^3.14.0"
django-filter = "^23.1"
djangorestframework-simplejwt = "^5.2.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
